+++ test/config_test.exs
@@ -4,85 +4,84 @@ defmodule Anthropic.ConfigTest do
 
   alias Anthropic.Config
 
-  test "opts/0 returns current opts" do
-    assert %Anthropic.Config{} = Config.opts()
+  describe "create/0 " do
+    test "returns current opts" do
+      assert %Anthropic.Config{} = Config.create([])
+    end
+
+    test "Environment takes precedence to default" do
+      assert %Anthropic.Config{api_key: "Loaded_for_tests"} = Config.create([])
+      Application.put_env(:anthropic, :api_key, "application.put_env.api_key")
+      assert %Anthropic.Config{api_key: "application.put_env.api_key"} = Config.create([])
+
+      assert %Anthropic.Config{api_key: "just_in_api_key"} =
+               Config.create(api_key: "just_in_api_key")
+    end
   end
 
   describe "validations" do
+    test "api_key" do
+      assert Anthropic.Config.validate_config(api_key: "good") == [api_key: "good"]
 
-    @tag :capture_log
-    test "top_p" do
-      Process.flag :trap_exit, true
-      {:ok, config} = GenServer.start_link(Anthropic.Config, [], name: :new)
-
-      catch_exit do
-        GenServer.call(config, {:reset, [top_p: 1.1]})
-      end
-
-      assert_received({:EXIT, ^config, {%ArgumentError{message: "Invalid top_p value, must be a float between 0.0 and 1.0."}, _}})
+      assert_raise(ArgumentError, ":api_key must be a String.t(). Got: nil", fn ->
+        Anthropic.Config.validate_config(api_key: nil)
+      end)
     end
 
-    test "valid top_p" do
-      assert %Anthropic.Config{top_p: 0.9} = Anthropic.Config.reset(top_p: 0.9)
+    test "api_url" do
+      assert_raise(ArgumentError, ":api_url must be a String.t(). Got: nil", fn ->
+        Anthropic.Config.validate_config(api_url: nil)
+      end)
     end
 
-    @tag :capture_log
     test "max_tokens" do
-      Process.flag :trap_exit, true
-      {:ok, config} = GenServer.start_link(Anthropic.Config, [], name: :new)
+      assert Anthropic.Config.validate_config(max_tokens: 10) == [max_tokens: 10]
 
-      catch_exit do
-        GenServer.call(config, {:reset, [max_tokens: -10]})
-      end
+      assert_raise(ArgumentError, "Invalid max_tokens value, must be a positive integer.", fn ->
+        Anthropic.Config.validate_config(max_tokens: "10")
+      end)
 
-      assert_received({:EXIT, ^config, {%ArgumentError{message: "Invalid max_tokens value, must be a positive integer."}, _}})
+      assert_raise(ArgumentError, "Invalid max_tokens value, must be a positive integer.", fn ->
+        Anthropic.Config.validate_config(max_tokens: -10)
+      end)
     end
 
-    @tag :capture_log
     test "temperature" do
-      Process.flag :trap_exit, true
-      {:ok, config} = GenServer.start_link(Anthropic.Config, [], name: :new)
+      assert Anthropic.Config.validate_config(temperature: 0.5) == [temperature: 0.5]
 
-      catch_exit do
-        GenServer.call(config, {:reset, [temperature: 1.1]})
-      end
+      assert_raise(
+        ArgumentError,
+        "Invalid temperature value, must be a float between 0.0 and 1.0.",
+        fn -> Anthropic.Config.validate_config(temperature: "10") end
+      )
 
-      assert_received({:EXIT, ^config, {%ArgumentError{message: "Invalid temperature value, must be a float between 0.0 and 1.0."}, _}})
-    end
-  end
-
-  describe "reset/1" do
-    test "will rewrite runtime options" do
-      Config.reset(temperature: 0.3)
-      assert 0.3 == Config.opts().temperature
-      Config.reset(temperature: 1.0)
+      assert_raise(
+        ArgumentError,
+        "Invalid temperature value, must be a float between 0.0 and 1.0.",
+        fn -> Anthropic.Config.validate_config(temperature: -10) end
+      )
     end
 
-    test "will rewrite api_key runtime options" do
-      Config.reset(api_key: "some other key")
-      assert "some other key" == Config.opts().api_key
-      Config.reset(api_key: "Loaded_for_tests")
+    test "top_p" do
+      assert Anthropic.Config.validate_config(top_p: 0.5) == [top_p: 0.5]
+
+      assert_raise(
+        ArgumentError,
+        "Invalid top_p value, must be a float between 0.0 and 1.0.",
+        fn -> Anthropic.Config.validate_config(top_p: "10") end
+      )
+
+      assert_raise(
+        ArgumentError,
+        "Invalid top_p value, must be a float between 0.0 and 1.0.",
+        fn -> Anthropic.Config.validate_config(top_p: -10) end
+      )
     end
-
-    # @tag :capture_log
-    # test "will raise on invalid options" do
-    #   Process.flag :trap_exit, true
-
-    #   catch_exit do
-    #     Config.reset(temperature: :hot)
-    #   end
-
-    #   Config.reset(temperature: :hot)
-
-    # assert_receive({:DOWN, _, _},100)
-    # assert_received({:EXIT, _, {%ArgumentError{message: "Invalid temperature value, must be a float between 0.0 and 1.0"}, _}})
-
-    # end
   end
 
   describe "Enumerable implementation" do
     test "count" do
-      assert Enum.count(%Anthropic.Config{}) == 8
+      assert Enum.count(%Anthropic.Config{}) == 3
     end
 
     test "member?" do
@@ -90,14 +89,16 @@ defmodule Anthropic.ConfigTest do
     end
 
     test "reduce function correctly handles :halted" do
-      config = %Config{model: "test-model", max_tokens: 500, temperature: 0.5}
+      config = %Config{api_key: "test-model", api_url: "url"}
       reducer = fn _x, acc -> {:cont, [(& &1) | acc]} end
+
       assert [] == Enumerable.reduce(config, {:halt, []}, reducer) |> elem(1) |> :lists.reverse()
 
       assert [] ==
                Enumerable.reduce(config, {:suspend, []}, reducer) |> elem(1) |> :lists.reverse()
 
       assert {:error, Enumerable.Anthropic.Config} == Enumerable.slice(config)
+      assert Enum.reduce(config, fn _e, acc -> acc end) == {:api_key, "test-model"}
     end
   end
 end